Create validation directory inside utils, move files
